etl/python/requirements.txt
#--- REQUIREMENTS TXT

lxml
psycopg2-binary
